Check for git-wip, not running it, and create buf-local autocmds
The `run 'git wip'` approach was, by far, the single longest thing vim
did when starting up -- and given my vimrc, that's fairly impressive.

Here we check for the presence of the `git-wip` executable, and take
advantage of vim-fugitive such that we only try to do a wip save for
buffers actually in a git repo.